requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flasgger==0.9.7.1
flask-cors==4.0.0
pytest==7.4.0
pytest-flask==1.2.0
flask-migrate==4.0.0
pydantic==2.11.7